Added Loadable interface for Data-API to allow using initializers. Added timeout-API to PlayerInfo. Fixed /kit command. Fixes #1412

+++ src/main/java/com/forgeessentials/util/FunctionHelper.java
@@ -332,17 +332,17 @@ public final class FunctionHelper {
 	 * @param timeInSec
 	 * @return Time in string format
 	 */
-	public static String parseTime(int timeInSec)
+	public static String parseTime(long timeInSec)
 	{
 		String uptime = "";
-		int weeks = timeInSec / (86400 * 7);
-		int remainder = timeInSec % (86400 * 7);
-		int days = remainder / 86400;
+		long weeks = timeInSec / (86400 * 7);
+		long remainder = timeInSec % (86400 * 7);
+		long days = remainder / 86400;
 		remainder = timeInSec % 86400;
-		int hours = remainder / 3600;
+		long hours = remainder / 3600;
 		remainder = timeInSec % 3600;
-		int minutes = remainder / 60;
-		int seconds = remainder % 60;
+		long minutes = remainder / 60;
+		long seconds = remainder % 60;
 
 		if (weeks != 0)
 		{
